pyproject.toml
[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "cookiecutter-svdvoort"
description = "Cookiecutter package"
version = "0.1.0"
readme = "README.md"
authors = ["svdvoort"]

license = "MIT"


[tool.poetry.dependencies]
python = ">=3.9,<3.12"

cookiecutter = "^2.6"
jinja2-git = "^1.3"
lice = "^0.6"
gitpython = "^3.1.43"
cruft = "^2.15.0"
toml = "^0.10.2"


[tool.poetry.group.test.dependencies]

flake8-pytest-style = "^2.0"

pytest = "^8.2"
pytest-randomly = "^3.15"
pytest-cookies = "^0.7"

safety = "^3.2"
tomli = "^2.0"

[tool.isort]
line_length = 100
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
multi_line_output = 3
use_parentheses = true
skip_glob = ["*/setup.py"]
filter_files = true

[tool.darglint]
docstring_style = "google"


[tool.pytest.ini_options]
norecursedirs = "{{project_name}} hooks *.egg .eggs dist build docs .tox .git __pycache__"
# Directories that are not visited by pytest collector:
# Extra options:
addopts = "--strict --tb=short --doctest-modules"
